+++ lib/grape_entity.rb
@@ -1,6 +1,6 @@
-require "grape/version"
+require "grape_entity/version"
 
 module Grape
-  autoload :Entity, 'grape/entity'
+  autoload :Entity, 'grape_entity/entity'
 end
