+++ lib/grape_entity/entity.rb
