+++ spec/grape_entity/entity_spec.rb
